pyproject.toml
[project]
name = "epistole"
version = "2.0.0"
description = "IMAP/SMTP email MCP server with local semantic search"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11,<3.14"
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-dotenv>=1.0.0",
    "chromadb>=0.5.0",
    "model2vec>=0.8.0",
    "pypdf>=5.0.0",
    "aiosqlite>=0.20.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24"]

[project.scripts]
epistole = "epistole.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python-preference = "system"
